Rearrange all plugins to the plugins directory in Calibration/HcalAlCaRecoProducers
Small changes

Code check

small fix

code check

Code check

make track_ instead of Track_
