--- Calibration/HcalAlCaRecoProducers/BuildFile.xml
@@ -1,30 +1,0 @@
-<use name="Calibration/IsolatedParticles"/>
-<use name="CommonTools/UtilAlgos"/>
-<use name="CondFormats/DataRecord"/>
-<use name="CondFormats/HcalObjects"/>
-<use name="DataFormats/CaloTowers"/>
-<use name="DataFormats/DetId"/>
-<use name="DataFormats/EcalDetId"/>
-<use name="DataFormats/EcalRecHit"/>
-<use name="DataFormats/EgammaReco"/>
-<use name="DataFormats/FEDRawData"/>
-<use name="DataFormats/HcalCalibObjects"/>
-<use name="DataFormats/HcalIsolatedTrack"/>
-<use name="DataFormats/HcalRecHit"/>
-<use name="DataFormats/JetReco"/>
-<use name="DataFormats/Luminosity"/>
-<use name="DataFormats/METReco"/>
-<use name="DataFormats/RecoCandidate"/>
-<use name="HLTrigger/HLTcore"/>
-<use name="FWCore/Framework"/>
-<use name="FWCore/ParameterSet"/>
-<use name="FWCore/PluginManager"/>
-<use name="RecoLocalCalo/EcalRecAlgos"/>
-<use name="RecoLocalCalo/HcalRecAlgos"/>
-<use name="TrackPropagation/SteppingHelixPropagator"/>
-<use name="DataFormats/Scalers"/>
-<use name="boost"/>
-<use name="hls"/>
-<use name="clhep"/>
-<use name="root"/>
-<flags EDM_PLUGIN="1"/>
