--- Calibration/HcalAlCaRecoProducers/interface/AlCaEcalHcalReadoutsProducer.h
@@ -1,51 +1,0 @@
-// -*- C++ -*-
-
-// system include files
-#include <memory>
-#include <string>
-// user include files
-#include "FWCore/Framework/interface/Frameworkfwd.h"
-#include "FWCore/Framework/interface/global/EDProducer.h"
-
-#include "FWCore/Framework/interface/Event.h"
-#include "FWCore/Framework/interface/MakerMacros.h"
-
-#include "FWCore/ParameterSet/interface/ParameterSet.h"
-#include "FWCore/Framework/interface/Frameworkfwd.h"
-#include "FWCore/Framework/interface/Event.h"
-#include "FWCore/Framework/interface/MakerMacros.h"
-#include "FWCore/Framework/interface/EventSetup.h"
-#include "FWCore/ParameterSet/interface/ParameterSet.h"
-
-#include "DataFormats/Common/interface/Ref.h"
-#include "DataFormats/CaloTowers/interface/CaloTowerCollection.h"
-#include "DataFormats/DetId/interface/DetId.h"
-
-#include "Geometry/Records/interface/IdealGeometryRecord.h"
-#include "Geometry/CaloGeometry/interface/CaloGeometry.h"
-#include "DataFormats/GeometryVector/interface/GlobalPoint.h"
-#include "DataFormats/CaloTowers/interface/CaloTowerDetId.h"
-#include "DataFormats/HcalRecHit/interface/HcalRecHitCollections.h"
-#include "DataFormats/EcalRecHit/interface/EcalRecHitCollections.h"
-
-//
-// class decleration
-//
-
-class AlCaEcalHcalReadoutsProducer : public edm::global::EDProducer<> {
-public:
-  explicit AlCaEcalHcalReadoutsProducer(const edm::ParameterSet &);
-
-  void produce(edm::StreamID, edm::Event &, const edm::EventSetup &) const override;
-
-private:
-  // ----------member data ---------------------------
-
-  edm::EDGetTokenT<HBHERecHitCollection> tok_hbhe_;
-  edm::EDGetTokenT<HORecHitCollection> tok_ho_;
-  edm::EDGetTokenT<HFRecHitCollection> tok_hf_;
-
-  edm::EDPutTokenT<HBHERecHitCollection> put_hbhe_;
-  edm::EDPutTokenT<HORecHitCollection> put_ho_;
-  edm::EDPutTokenT<HFRecHitCollection> put_hf_;
-};
