+++ Calibration/HcalAlCaRecoProducers/plugins/AlCaGammaJetProducer.cc
