+++ Calibration/HcalAlCaRecoProducers/plugins/BuildFile.xml
@@ -1,10 +1,33 @@
+<use name="Calibration/IsolatedParticles"/>
+<use name="CommonTools/UtilAlgos"/>
+<use name="CondFormats/DataRecord"/>
+<use name="CondFormats/HcalObjects"/>
+<use name="DataFormats/CaloTowers"/>
+<use name="DataFormats/DetId"/>
+<use name="DataFormats/EcalDetId"/>
+<use name="DataFormats/EcalRecHit"/>
+<use name="DataFormats/EgammaReco"/>
+<use name="DataFormats/FEDRawData"/>
+<use name="DataFormats/HcalCalibObjects"/>
+<use name="DataFormats/HcalIsolatedTrack"/>
+<use name="DataFormats/HcalRecHit"/>
+<use name="DataFormats/JetReco"/>
+<use name="DataFormats/Luminosity"/>
+<use name="DataFormats/METReco"/>
+<use name="DataFormats/RecoCandidate"/>
+<use name="DataFormats/Scalers"/>
 <use name="FWCore/Framework"/>
 <use name="FWCore/MessageLogger"/>
 <use name="FWCore/ParameterSet"/>
 <use name="HLTrigger/HLTcore"/>
-<use name="Calibration/IsolatedParticles"/>
-<use name="DataFormats/JetReco"/>
+<use name="RecoLocalCalo/EcalRecAlgos"/>
+<use name="RecoLocalCalo/HcalRecAlgos"/>
+<use name="TrackPropagation/SteppingHelixPropagator"/>
 <use name="rootphysics"/>
+<use name="boost"/>
+<use name="hls"/>
+<use name="clhep"/>
+<use name="root"/>
 <library file="*.cc" name="CalibrationHcalAlCaRecoProducersPlugins">
   <flags EDM_PLUGIN="1"/>
 </library>
