--- Calibration/HcalAlCaRecoProducers/src/SealModule.cc
@@ -1,9 +1,0 @@
-#include "FWCore/PluginManager/interface/ModuleDef.h"
-#include "FWCore/Framework/interface/MakerMacros.h"
-#include "Calibration/HcalAlCaRecoProducers/interface/AlCaEcalHcalReadoutsProducer.h"
-#include "Calibration/HcalAlCaRecoProducers/src/ProducerAnalyzer.h"
-
-using cms::ProducerAnalyzer;
-
-DEFINE_FWK_MODULE(AlCaEcalHcalReadoutsProducer);
-DEFINE_FWK_MODULE(ProducerAnalyzer);
